backend/requirements.txt
# FELICITAFAC - Sistema de Facturación Electrónica para Perú
# Dependencias optimizadas para MySQL y hosting compartido
# Compatible con Python 3.12+

# Framework principal
Django>=4.2,<5.0
djangorestframework>=3.14.0

# Base de datos MySQL
mysqlclient>=2.1.0
PyMySQL>=1.0.0

# Autenticación y seguridad
djangorestframework-simplejwt>=5.0.0
django-cors-headers>=4.0.0
cryptography>=40.0.0

# Configuración y variables de entorno
python-decouple>=3.6
python-dotenv>=1.0.0

# APIs y requests
requests>=2.28.0
urllib3>=1.26.0

# Filtros y búsqueda
django-filter>=22.0

# Fechas y tiempo
python-dateutil>=2.8.0
pytz>=2023.3

# Reportes y PDF
reportlab>=4.0.0
openpyxl>=3.1.0
xlsxwriter>=3.0.0

# Manejo de archivos
Pillow>=10.0.0

# Desarrollo y testing
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0

# Calidad de código
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0

# Utilidades básicas
setuptools>=65.0.0
wheel>=0.40.0

# Logging y monitoring
django-extensions>=3.2.0

# Integración SUNAT Perú
xmltodict>=0.13.0
lxml>=4.9.0

# Cache y performance
redis>=4.0.0
django-redis>=5.0.0

# Hosting compartido
gunicorn>=21.0.0
whitenoise>=6.0.0

# Formatos de fecha en español
Babel>=2.12.0

# Generación de códigos QR para facturas
qrcode[pil]>=7.4.0

# Validación de datos
validators>=0.20.0

# Compresión de respuestas
django-compressor>=4.0

# Manejo de archivos estáticos
django-storages>=1.14.0